backend/requirements.txt
asgiref==3.11.1
Django==5.2.11
django-cors-headers==4.9.0
djangorestframework==3.16.1
gunicorn==25.0.3
mysqlclient==2.2.7
packaging==26.0
sqlparse==0.5.5
